add folder support & nvim config
sourced from my old configs & https://www.youtube.com/watch?v=w7i4amO_zaE

+++ setup.py
@@ -2,26 +2,32 @@
 
 import os
 
-def add(config_dir, target_dir):
+def add(config_dir, target_dir, dry_run):
     """
     Creates `target_dir/* -> config_dir/*` symlinks, deleting existing targets.
     """
     if not os.path.exists(config_dir):
         raise ValueError(f'Error: config dir {config_dir} does not exist')
     elif not os.path.exists(target_dir):
-        raise ValueError(f'Error: target dir {target_dir} does not exist')
+        os.mkdir(target_dir)
 
     for config in os.listdir(config_dir):
         config_path = f'{config_dir}/{config}'
         target_path = f'{target_dir}/{config}'
-        print(f'{target_path} -> {config_path}')
-        try:
-            os.symlink(config_path, target_path)
-        except FileExistsError:
-            os.unlink(target_path)
-            os.symlink(config_path, target_path)
+        
+        if os.path.isdir(config_path):
+            add(config_path, target_path, dry_run)
+        else:
+            print(f'{target_path} -> {config_path}')
+            if dry_run:
+                continue
+            try:
+                os.symlink(config_path, target_path)
+            except FileExistsError:
+                os.unlink(target_path)
+                os.symlink(config_path, target_path)
 
-def remove(config_dir, target_dir):
+def remove(config_dir, target_dir, dry_run):
     """
     Removes the targets for a given config.
     """
@@ -31,12 +37,19 @@ def remove(config_dir, target_dir):
         raise ValueError(f'Error: target dir {target_dir} does not exist')
 
     for config in os.listdir(config_dir):
+        config_path = f'{config_dir}/{config}'
         target_path = f'{target_dir}/{config}'
-        print(f'Removing {target_path}')
-        try:
-            os.unlink(target_path)
-        except FileNotFoundError:
-            pass
+
+        if os.path.isdir(target_path):
+            remove(config_path, target_path, dry_run)
+        else:
+            print(f'Removing {target_path}')
+            if dry_run:
+                continue
+            try:
+                os.unlink(target_path)
+            except FileNotFoundError:
+                pass
 
 if __name__ == '__main__':
     current_dir = os.path.dirname(os.path.realpath(__file__))
@@ -44,6 +57,16 @@ if __name__ == '__main__':
     
     zsh_config = os.path.join(current_dir, 'zsh')
     zsh_target = home_dir
-    add(zsh_config, zsh_target)
+    add(zsh_config, zsh_target, True)
+
+    nvim_config = os.path.join(current_dir, 'nvim')
+    nvim_target = home_dir
+    add(nvim_config, nvim_target, False)
+    remove(nvim_config, nvim_target, True)
+
+    foo_config = os.path.join(current_dir, 'foo')
+    foo_target = home_dir
+    add(foo_config, foo_target, False)
+    remove(foo_config, foo_target, True)
